Removed yargs in favour of node-getopt-long for better more descriptive arguments
