+++ package.json
@@ -75,7 +75,6 @@
     "through": "~2.3.6",
     "through2": "~0.6.3",
     "vinyl": "~0.4.6",
-    "vinyl-sourcemaps-apply": "^0.2.0",
-    "yargs": "^3.26.0"
+    "vinyl-sourcemaps-apply": "^0.2.0"
   }
 }
